Added exception and backtrace as a comment at the bottom of the HTML
I frequently use curl testing in Rails, but when better_errors intercepts
an exception on those requests I get a large output that is unusable at
the command line. I simply added an HTML comment to the end of the HTML
that includes the exception and backtrace.

This shouldn't effect viewing the page in the browser at all, but makes
life at the command line much easier when tracking down errors.

Example output, trimmed for brevity: https://gist.github.com/4347455