Allowing any remote host to get better_errors
I'm indicating a simpler way to allow better_errors be used in heterogeneous development without the need to get the IP environments indicating that your development environment is using. 

At first I was worried that this could inadvertently lead to gem to be used in a production environment, however this line I noticed that the gem already has a [mechanism][0] to prevent this from occurring and to supplant it needs to let the developer explicitly enabled in config. 

Hope that helps.

[0]: https://github.com/charliesome/better_errors/blob/master/lib/better_errors/rails.rb#L21