Spec to cover REPL::Pry eval
This required changing the method of loading REPL::Pry. Since it can only be included when Pry is in the bundle, REPL attempts to load it only during REPL provider selection. In order to test it effectively, we must load it only when needed for a test, and then unload it afterward so that the rest of the spec examples aren't polluted by the existence of the constant.

This is all very hacky, and would be better refactored into a separate gem such as `better_errors-pry`.