Allow skipping variable inspection by class name
This adds a new global setting `BetterErrors.ignored_classes` that will
skip inspecting the value of local and instance variables of this type.
The default setting ignores action dispatch request and response
instances, which are usually not printed anyways due to exceeding the
limit of `BetterErrors.maximum_variable_inspect_size`, but also avoids
all the memory allocations and database queries that are triggered by
calling `#inspect` on variables.

The default setting was chosen to avoid excessive database queries for
unloaded relations, because both the `ActionDispatch::Request` and the
`ActionDispatch::Response` objects hold many references to the
controller instance, causing the same instance variables to be inspected
over and over again.